requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.5
Django==3.2.7
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-js-asset==1.2.2
django-social-share==2.2.1
idna==3.2
Pillow==8.3.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.6
